pyproject.toml
[project]
name = "mcpdemolangchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=0.1.0",
    "langchain-mcp-adapters>=0.1.7",
    "langgraph>=0.4.8",
    "mcp>=1.9.4",
    "beautifulsoup4>=4.13.4",
    "openai==1.58.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "streamlit==1.41.1",
    "tavily-python>=0.7.8",
    "agno>=1.7.0",
    "yfinance>=0.2.64",
    "duckduckgo-search>=8.0.4",
    "pandas>=2.3.0",
    "duckdb==1.1.3",
    "phidata>=2.7.10",
    "numpy==1.26.4",
]
